little reorg of the consensus manager file (#1861)
* little reorg of the consensus manager file

* post PR changes

* Fixes as per review
