+++ src/Stratis.Bitcoin/Consensus/Validators/BlockValidator.cs
@@ -3,6 +3,7 @@ using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 using NBitcoin;
 using Stratis.Bitcoin.Consensus.Rules;
+using Stratis.Bitcoin.Consensus.ValidationResults;
 using Stratis.Bitcoin.Primitives;
 using Stratis.Bitcoin.Utilities;
